(ASC-1393) reuse gen_dict_extract function
Looking up for rpc_product_release has been changed recently in
commit 46760db71ee69387866a3cd0aa649ebb857630c3. The old way of
getting 'epc_product_release' is no longer working so it will
break other system tests (test_dhcp_agents.py and test_swift_stat.py).

This PR to relocate the method to utils.py file so other tests can
utilize it. Related test cases are also modified in this PR
