+++ molecule/default/tests/utils.py
@@ -127,3 +127,27 @@ def get_expected_value(resource_type,
     print(show_all_status(run_on_host))
 
     return False
+
+
+def gen_dict_extract(key, var):
+    """Produce a generator that can iterate over all found values for a given
+    key in a given dictionary
+
+    Args:
+        key(str): key to lookup
+        var(dict): dictionary to recursively search
+
+    Returns: generator
+    """
+
+    if hasattr(var, 'iteritems'):
+        for k, v in var.iteritems():
+            if k == key:
+                yield v
+            if isinstance(v, dict):
+                for result in gen_dict_extract(key, v):
+                    yield result
+            elif isinstance(v, list):
+                for d in v:
+                    for result in gen_dict_extract(key, d):
+                        yield result
